requirements.txt
google-generativeai==0.1.0
python-dotenv==0.20.0
argparse==1.4.0
